Domain Check ============
It is a fact universally acknowledged that a user in
possession of a browser must be in want of a feature like
warnings about nefarious sites, ad block, or warnings
before accidentally visiting saucy content. Of course,
one way of providing such a service involves sending
every URL the user visits to a central server in order
to check it against such lists. Unfortunately, such a
solution entirely sacrifices the user's privacy, by
offering the operator of the central server a complete
list of every site the user visits.

A privacy preserving solution to this problem requires just
hosting a regularly compiled database of domains so that
the user's device, not a central server, performs the
checking. This is a good solution, but it comes with a
cost: storage and CPU. Many of these lists contain
hundreds of thousands or millions of entries, and mobile
devices often have slow storage. As a result, simply
shipping a list of domain names (even one with a generic
compression like Zip) results in big downloads, big
stored files and slow lookup times.

In this puzzle you will create two programs (or one program
which performs the two different tasks controlled by a
command line flag): encoding a raw list of domains into
a compressed database, and, given the database produced
in the first step, looking up a domain name to see if it
is in the list.

So, for example, to compress your program might be invoked
like:

    java compress domains.txt domains.dat

Compression
-----------

The data to compress consists only of a list of domain
names, one per line, separated by the new line character
(`\n`). An example file containing about a hundred of
the most popular domain names is included in this folder
as `domains.txt`. There's also 100,000 and 500,000 in
`domains-more.txt` and `domains-tons.txt`.

Even the 7.7MB of the list with 500k domains is small
compared to some lists, which can reach into the hundreds
of MB in production.

Important Notes
---------------

We've intentionally trimmed these lists and include just
.com, .net and .org domain names. In reality there are
hundreds of different top level domains. There is a well
maintained list (with libraries in many languages) called
the public suffix list, but there are pros and cons to
requiring it in your solution.

In particular, requiring the PSL or other data which is
hardcoded in the binary may decrease the size of the
database, but will do so at the cost of increasing the
binary size. We do not intend this to be a puzzle in
compiling small binaries, but at the same time we would
not expect to see tons of static data related to the
problem compiled into the binary.

Your program should be capable of taking any input list
of domains and outputting a valid database. That having
been said, we will not run it with inputs containing
more than 10 million domain names.

We not only accept, but encourage the use of good
off-the-shelf libraries for things like Bloom filters
(and friends), indexes, various tree structures, hashes,
etc.

A probabilistic or heuristic approach which produces the
correct answer with high confidence but fails once in a
while may be a good answer. We're interested in the
trade-offs and design decisions which you make in
solving this puzzle.

List Attribution
----------------

The domain lists used for this puzzle were compiled from
the "Majestic Million" domain list.